# app/backend/api-gateway/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Task Tracker API
  version: "1.0.0"
  description: |
    External REST API exposed by the api-gateway.

    **Internal** service-to-service communication (api-gateway → auth-service,
    api-gateway → project-service) uses gRPC and is defined separately in
    `auth-service/api/proto/auth_service.proto` and
    `project-service/api/proto/project_service.proto`.
    Those contracts are intentionally NOT duplicated here.
servers:
  - url: http://localhost:8000
    description: Local development (via api-gateway)

# ── Security ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:          { type: integer }
        manager_id:  { type: integer }
        name:        { type: string }
        description: { type: string }
        status:      { type: string, enum: [PROJECT_STATUS_IN_WORK, PROJECT_STATUS_ENDED] }
        start_date:  { type: string, format: date-time }
        end_date:    { type: string, format: date-time, nullable: true }
    Task:
      type: object
      properties:
        id:          { type: integer }
        project_id:  { type: integer }
        assignee_id: { type: integer, nullable: true }
        name:        { type: string }
        description: { type: string, nullable: true }
        priority:    { type: string, enum: [TASK_PRIORITY_LOW, TASK_PRIORITY_MEDIUM, TASK_PRIORITY_HIGH] }
        difficulty:  { type: string, enum: [TASK_DIFFICULTY_EASY, TASK_DIFFICULTY_MEDIUM, TASK_DIFFICULTY_HARD] }
        status:      { type: string, enum: [TASK_STATUS_NOT_STARTED, TASK_STATUS_IN_WORK, TASK_STATUS_ON_REVIEW, TASK_STATUS_COMPLETED] }
        start_date:  { type: string, format: date-time, nullable: true }
        end_date:    { type: string, format: date-time, nullable: true }
    Comment:
      type: object
      properties:
        id:            { type: integer }
        author_id:     { type: integer }
        task_id:       { type: integer }
        content:       { type: string }
        creation_date: { type: string, format: date-time }
...
    OperationResponse:
      type: object
      properties:
        message: { type: string }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/admin/projects/get:
    get:
      tags: [Admin]
      summary: Get a project by ID or name
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /api/admin/projects:
    put:
      tags: [Admin]
      summary: Update a project
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
    delete:
      tags: [Admin]
      summary: Delete a project
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
  /api/admin/tasks/get:
    get:
      tags: [Admin]
      summary: Get a task by ID, project or name
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /api/admin/tasks:
    put:
      tags: [Admin]
      summary: Update a task
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
    delete:
      tags: [Admin]
      summary: Delete a task
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
  /api/admin/comments/get:
    get:
      tags: [Admin]
      summary: Get a comment by ID, author or task
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /api/admin/comments:
    put:
      tags: [Admin]
      summary: Update a comment
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
    delete:
      tags: [Admin]
      summary: Delete a comment
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }